requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
web3==6.15.1
requests==2.31.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9 